Restore monolithic server.js: Remove all organized folders and restore original code exactly as it was
