--- routes/api.js
@@ -1,34 +1,0 @@
-const express = require('express');
-const { BuyOrder, BannedUser, generateOrderId } = require('../models');
-const { adminIds } = require('../config');
-
-const router = express.Router();
-
-// Wallet Address Endpoint
-router.get('/get-wallet-address', (req, res) => {
-    try {
-        const walletAddress = process.env.WALLET_ADDRESS;
-        
-        if (!walletAddress) {
-            return res.status(500).json({
-                success: false,
-                error: 'Wallet address not configured'
-            });
-        }
-
-        res.json({
-            success: true,
-            walletAddress: walletAddress
-        });
-    } catch (error) {
-        console.error('Error getting wallet address:', error);
-        res.status(500).json({
-            success: false,
-            error: 'Internal server error'
-        });
-    }
-});
-
-
-
-module.exports = router;
